model/requirements.txt
fastapi
pydantic
google.generativeai
dotenv